requirements.txt

python-dotenv
sendgrid==6.6.0